Made it possible to use previously created certs
There is a problem when we want to include previously created
certificates into the truststore.
Since they could not be there.

To solve this I added a test to check if preexisting certificates
are there and if so that they can also be used.

Further I added a wait configuration to be able to wait for bob process
which run in parallel.
This allows to run multiple docker instances with bob without defining
a run order.
This is a not the best solution but a fast one.